% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteusSILAC.R
\docType{data}
\name{evi}
\alias{evi}
\title{Evidence data}
\format{A data frame with 110,502 rows and 10 columns}
\description{
Evidence data from a SILAC proteomics experiment in 2 time points, 3 replicates each.
}
