% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteusSILAC.R
\docType{data}
\name{meta}
\alias{meta}
\title{Proteomics metadata}
\format{A \code{data.frame} with 6 rows and 5 columns (experiment, measure, sample, replicate).}
\description{
Metadata describing a SILAC proteomics experiment at 2 time points with 3 replicates each.
}
