% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteusSILAC.R
\docType{data}
\name{proteusSILAC}
\alias{proteusSILAC}
\title{Proteus example SILAC data}
\format{Four objects from different stages of data processing:
\describe{
  \item{\code{evi}}{Pre-processed evidence data (data frame)}
  \item{\code{meta}}{Metadata (data frame)}
  \item{\code{pepdat}}{Peptide data (\code{proteusData} object)}
  \item{\code{prodat}}{Protein data (\code{proteusData} object)}
}

This package also contains metadata file (see examples below). The original evidence file is not included due to large size.}
\description{
Example SILAC data set for \code{proteus} package. It is a small subset of
OHT-vs-VEH comparison from Ly et al. 2018 (https://wellcomeopenresearch.org/articles/3-51/).
This data set contains two time points in three replicates each.
}
\examples{
library(proteusSILAC)
data(proteusSILAC)
meta
head(pepdat$tab)
head(prodat$tab)

metadataFile <- system.file("extdata", "metadata.txt", package="proteusSILAC")
}
